Convert Workload CVEs deployment page to TbodyUnified
