+++ ui/apps/platform/src/Containers/Vulnerabilities/WorkloadCves/Tables/DeploymentVulnerabilitiesTable.tsx
@@ -2,29 +2,24 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import { ExpandableRowContent, Table, Tbody, Td, Th, Thead, Tr } from '@patternfly/react-table';
 import { gql } from '@apollo/client';
-import { min } from 'date-fns';
 
 import useSet from 'hooks/useSet';
 import { UseURLSortResult } from 'hooks/useURLSort';
 import VulnerabilitySeverityIconText from 'Components/PatternFly/IconText/VulnerabilitySeverityIconText';
-import { VulnerabilitySeverity, VulnerabilityState } from 'types/cve.proto';
+import { VulnerabilityState } from 'types/cve.proto';
 import VulnerabilityFixableIconText from 'Components/PatternFly/IconText/VulnerabilityFixableIconText';
 import { DynamicColumnIcon } from 'Components/DynamicIcon';
 import DateDistance from 'Components/DateDistance';
+import TbodyUnified from 'Components/TableStateTemplates/TbodyUnified';
+import { TableUIState } from 'utils/getTableUIState';
 import { getWorkloadEntityPagePath } from '../../utils/searchUtils';
 
-import EmptyTableResults from '../components/EmptyTableResults';
 import DeploymentComponentVulnerabilitiesTable, {
-    DeploymentComponentVulnerability,
-    ImageMetadataContext,
     deploymentComponentVulnerabilitiesFragment,
 } from './DeploymentComponentVulnerabilitiesTable';
-import {
-    getIsSomeVulnerabilityFixable,
-    getHighestVulnerabilitySeverity,
-} from '../../utils/vulnerabilityUtils';
 import PendingExceptionLabelLayout from '../components/PendingExceptionLabelLayout';
 import PartialCVEDataAlert from '../../components/PartialCVEDataAlert';
+import { FormattedDeploymentVulnerability } from './table.utils';
 
 export const deploymentWithVulnerabilitiesFragment = gql`
     ${deploymentComponentVulnerabilitiesFragment}
@@ -49,107 +44,22 @@ export const deploymentWithVulnerabilitiesFragment = gql`
     }
 `;
 
-export type DeploymentWithVulnerabilities = {
-    id: string;
-    images: ImageMetadataContext[];
-    imageVulnerabilities: {
-        vulnerabilityId: string;
-        cve: string;
-        operatingSystem: string;
-        summary: string;
-        pendingExceptionCount: number;
-        images: {
-            imageId: string;
-            imageComponents: DeploymentComponentVulnerability[];
-        }[];
-    }[];
-};
-
-type DeploymentVulnerabilityImageMapping = {
-    imageMetadataContext: ImageMetadataContext;
-    componentVulnerabilities: DeploymentComponentVulnerability[];
-};
-
-function formatVulnerabilityData(deployment: DeploymentWithVulnerabilities): {
-    vulnerabilityId: string;
-    cve: string;
-    operatingSystem: string;
-    severity: VulnerabilitySeverity;
-    isFixable: boolean;
-    discoveredAtImage: Date | null;
-    summary: string;
-    affectedComponentsText: string;
-    images: DeploymentVulnerabilityImageMapping[];
-    pendingExceptionCount: number;
-}[] {
-    // Create a map of image ID to image metadata for easy lookup
-    // We use 'Partial' here because there is no guarantee that the image will be found
-    const imageMap: Partial<Record<string, ImageMetadataContext>> = {};
-    deployment.images.forEach((image) => {
-        imageMap[image.id] = image;
-    });
-
-    return deployment.imageVulnerabilities.map((vulnerability) => {
-        const { vulnerabilityId, cve, operatingSystem, summary, images, pendingExceptionCount } =
-            vulnerability;
-        // Severity, Fixability, and Discovered date are all based on the aggregate value of all components
-        const allVulnerableComponents = vulnerability.images.flatMap((img) => img.imageComponents);
-        const allVulnerabilities = allVulnerableComponents.flatMap((c) => c.imageVulnerabilities);
-        const highestVulnSeverity = getHighestVulnerabilitySeverity(allVulnerabilities);
-        const isFixableInDeployment = getIsSomeVulnerabilityFixable(allVulnerabilities);
-        const allDiscoveredDates = allVulnerableComponents
-            .flatMap((c) => c.imageVulnerabilities.map((v) => v.discoveredAtImage))
-            .filter((d): d is string => d !== null);
-        const oldestDiscoveredVulnDate = min(...allDiscoveredDates);
-        // TODO This logic is used in many places, could extract to a util
-        const uniqueComponents = new Set(allVulnerableComponents.map((c) => c.name));
-        const affectedComponentsText =
-            uniqueComponents.size === 1
-                ? uniqueComponents.values().next().value
-                : `${uniqueComponents.size} components`;
-
-        const vulnerabilityImages = images
-            .map((img) => ({
-                imageMetadataContext: imageMap[img.imageId],
-                componentVulnerabilities: img.imageComponents,
-            }))
-            // filter out values where the vulnerability->image mapping is missing
-            .filter(
-                (vulnImageMap): vulnImageMap is DeploymentVulnerabilityImageMapping =>
-                    !!vulnImageMap.imageMetadataContext
-            );
-
-        return {
-            vulnerabilityId,
-            cve,
-            operatingSystem,
-            severity: highestVulnSeverity,
-            isFixable: isFixableInDeployment,
-            discoveredAtImage: oldestDiscoveredVulnDate,
-            summary,
-            affectedComponentsText,
-            images: vulnerabilityImages,
-            pendingExceptionCount,
-        };
-    });
-}
-
 export type DeploymentVulnerabilitiesTableProps = {
-    deployment: DeploymentWithVulnerabilities;
+    tableState: TableUIState<FormattedDeploymentVulnerability>;
     getSortParams: UseURLSortResult['getSortParams'];
     isFiltered: boolean;
     vulnerabilityState: VulnerabilityState | undefined; // TODO Make Required when the ROX_VULN_MGMT_UNIFIED_CVE_DEFERRAL feature flag is removed
+    onClearFilters: () => void;
 };
 
 function DeploymentVulnerabilitiesTable({
-    deployment,
+    tableState,
     getSortParams,
     isFiltered,
     vulnerabilityState,
+    onClearFilters,
 }: DeploymentVulnerabilitiesTableProps) {
     const expandedRowSet = useSet<string>();
-
-    const vulnerabilities = formatVulnerabilityData(deployment);
 
     return (
         <Table variant="compact">
@@ -170,85 +80,94 @@ function DeploymentVulnerabilitiesTable({
                     <Th>First discovered</Th>
                 </Tr>
             </Thead>
-            {vulnerabilities.length === 0 && <EmptyTableResults colSpan={7} />}
-            {vulnerabilities.map((vulnerability, rowIndex) => {
-                const {
-                    vulnerabilityId,
-                    cve,
-                    operatingSystem,
-                    severity,
-                    summary,
-                    isFixable,
-                    images,
-                    affectedComponentsText,
-                    discoveredAtImage,
-                    pendingExceptionCount,
-                } = vulnerability;
-                const isExpanded = expandedRowSet.has(vulnerabilityId);
+            <TbodyUnified
+                colSpan={7}
+                tableState={tableState}
+                emptyProps={{ message: 'There were no CVEs detected for this deployment' }}
+                filteredEmptyProps={{ onClearFilters }}
+                renderer={({ data }) =>
+                    data.map((vulnerability, rowIndex) => {
+                        const {
+                            vulnerabilityId,
+                            cve,
+                            operatingSystem,
+                            severity,
+                            summary,
+                            isFixable,
+                            images,
+                            affectedComponentsText,
+                            discoveredAtImage,
+                            pendingExceptionCount,
+                        } = vulnerability;
+                        const isExpanded = expandedRowSet.has(vulnerabilityId);
 
-                return (
-                    <Tbody key={vulnerabilityId} isExpanded={isExpanded}>
-                        <Tr>
-                            <Td
-                                expand={{
-                                    rowIndex,
-                                    isExpanded,
-                                    onToggle: () => expandedRowSet.toggle(vulnerabilityId),
-                                }}
-                            />
-                            <Td dataLabel="CVE" modifier="nowrap">
-                                <PendingExceptionLabelLayout
-                                    hasPendingException={pendingExceptionCount > 0}
-                                    cve={cve}
-                                    vulnerabilityState={vulnerabilityState}
-                                >
-                                    <Link
-                                        to={getWorkloadEntityPagePath(
-                                            'CVE',
-                                            cve,
-                                            vulnerabilityState
-                                        )}
-                                    >
-                                        {cve}
-                                    </Link>
-                                </PendingExceptionLabelLayout>
-                            </Td>
-                            <Td modifier="nowrap" dataLabel="OS">
-                                {operatingSystem}
-                            </Td>
-                            <Td modifier="nowrap" dataLabel="Severity">
-                                <VulnerabilitySeverityIconText severity={severity} />
-                            </Td>
-                            <Td modifier="nowrap" dataLabel="CVE Status">
-                                <VulnerabilityFixableIconText isFixable={isFixable} />
-                            </Td>
-                            <Td dataLabel="Affected components">{affectedComponentsText}</Td>
-                            <Td modifier="nowrap" dataLabel="First discovered">
-                                <DateDistance date={discoveredAtImage} />
-                            </Td>
-                        </Tr>
-                        <Tr isExpanded={isExpanded}>
-                            <Td />
-                            <Td colSpan={6}>
-                                <ExpandableRowContent>
-                                    {summary && images.length > 0 ? (
-                                        <>
-                                            <p className="pf-v5-u-mb-md">{summary}</p>
-                                            <DeploymentComponentVulnerabilitiesTable
-                                                images={images}
-                                                cve={cve}
-                                                vulnerabilityState={vulnerabilityState}
-                                            />
-                                        </>
-                                    ) : (
-                                        <PartialCVEDataAlert />
-                                    )}
-                                </ExpandableRowContent>
-                            </Td>
-                        </Tr>
-                    </Tbody>
-                );
-            })}
+                        return (
+                            <Tbody key={vulnerabilityId} isExpanded={isExpanded}>
+                                <Tr>
+                                    <Td
+                                        expand={{
+                                            rowIndex,
+                                            isExpanded,
+                                            onToggle: () => expandedRowSet.toggle(vulnerabilityId),
+                                        }}
+                                    />
+                                    <Td dataLabel="CVE" modifier="nowrap">
+                                        <PendingExceptionLabelLayout
+                                            hasPendingException={pendingExceptionCount > 0}
+                                            cve={cve}
+                                            vulnerabilityState={vulnerabilityState}
+                                        >
+                                            <Link
+                                                to={getWorkloadEntityPagePath(
+                                                    'CVE',
+                                                    cve,
+                                                    vulnerabilityState
+                                                )}
+                                            >
+                                                {cve}
+                                            </Link>
+                                        </PendingExceptionLabelLayout>
+                                    </Td>
+                                    <Td modifier="nowrap" dataLabel="OS">
+                                        {operatingSystem}
+                                    </Td>
+                                    <Td modifier="nowrap" dataLabel="Severity">
+                                        <VulnerabilitySeverityIconText severity={severity} />
+                                    </Td>
+                                    <Td modifier="nowrap" dataLabel="CVE Status">
+                                        <VulnerabilityFixableIconText isFixable={isFixable} />
+                                    </Td>
+                                    <Td dataLabel="Affected components">
+                                        {affectedComponentsText}
+                                    </Td>
+                                    <Td modifier="nowrap" dataLabel="First discovered">
+                                        <DateDistance date={discoveredAtImage} />
+                                    </Td>
+                                </Tr>
+                                <Tr isExpanded={isExpanded}>
+                                    <Td />
+                                    <Td colSpan={6}>
+                                        <ExpandableRowContent>
+                                            {summary && images.length > 0 ? (
+                                                <>
+                                                    <p className="pf-v5-u-mb-md">{summary}</p>
+                                                    <DeploymentComponentVulnerabilitiesTable
+                                                        images={images}
+                                                        cve={cve}
+                                                        vulnerabilityState={vulnerabilityState}
+                                                    />
+                                                </>
+                                            ) : (
+                                                <PartialCVEDataAlert />
+                                            )}
+                                        </ExpandableRowContent>
+                                    </Td>
+                                </Tr>
+                            </Tbody>
+                        );
+                    })
+                }
+            />
         </Table>
     );
 }
